LPD-35542 Move from defaultpermissions.kernel to kernel.defaultpermissions
