+++ modules/apps/portal/portal-default-permissions-web/src/main/resources/META-INF/resources/init.jsp
@@ -20,13 +20,13 @@ taglib uri="http://liferay.com/tld/util" prefix="liferay-util" %>
 <%@ page import="com.liferay.frontend.taglib.clay.servlet.taglib.util.DropdownItem" %><%@
 page import="com.liferay.petra.string.StringBundler" %><%@
 page import="com.liferay.petra.string.StringPool" %><%@
-page import="com.liferay.portal.defaultpermissions.kernel.configuration.manager.PortalDefaultPermissionsConfigurationManager" %><%@
 page import="com.liferay.portal.defaultpermissions.web.internal.constants.PortalDefaultPermissionsWebKeys" %><%@
 page import="com.liferay.portal.defaultpermissions.web.internal.display.context.BaseViewPortalDefaultPermissionsConfigurationDisplayContext" %><%@
 page import="com.liferay.portal.defaultpermissions.web.internal.display.context.EditPortalDefaultPermissionsConfigurationDisplayContext" %><%@
 page import="com.liferay.portal.defaultpermissions.web.internal.display.context.PortalDefaultPermissionsManagementToolbarDisplayContext" %><%@
 page import="com.liferay.portal.defaultpermissions.web.internal.portlet.action.EditPortalDefaultPermissionsConfigurationMVCActionCommand" %><%@
 page import="com.liferay.portal.kernel.dao.search.SearchContainer" %><%@
+page import="com.liferay.portal.kernel.defaultpermissions.configuration.manager.PortalDefaultPermissionsConfigurationManager" %><%@
 page import="com.liferay.portal.kernel.language.LanguageUtil" %><%@
 page import="com.liferay.portal.kernel.model.Role" %><%@
 page import="com.liferay.portal.kernel.model.role.RoleConstants" %><%@
