+++ portal-kernel/src/com/liferay/portal/kernel/defaultpermissions/configuration/manager/packageinfo
