+++ portal-kernel/src/com/liferay/portal/kernel/defaultpermissions/resource/PortalDefaultPermissionsModelResourceRegistryUtil.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.defaultpermissions.kernel.resource;
+package com.liferay.portal.kernel.defaultpermissions.resource;
 
 import com.liferay.portal.kernel.module.service.Snapshot;
