+++ portal-kernel/src/com/liferay/portal/kernel/defaultpermissions/resource/packageinfo
